fixing deleteByNim with token

+++ app/Http/Controllers/MahasiswasController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Mahasiswa;
+use Firebase\JWT\JWT;
 use Illuminate\Http\Request;
+use Firebase\JWT\Key;
 
 class MahasiswasController extends Controller
 {
@@ -34,16 +36,31 @@ class MahasiswasController extends Controller
         ]);
     }
 
-    public function deleteByNim(Request $request) {
-        $mahasiswa = Mahasiswa::find($request->nim);
+    public function deleteByNim(Request $request)
+    {
+        $headerToken = $request->header('token');
+        $jwtDecoded = JWT::decode($headerToken, new Key(env('JWT_SECRET'), 'HS256'));
+        $nimUser = $jwtDecoded->sub;
+
+        $mahasiswa = Mahasiswa::find($request->nim)->first();
+
+        if ($nimUser != $request->nim) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Delete tidak sesuai'
+            ]);
+        }
+
         $mahasiswa->delete();
         return response()->json([
             'status' => true,
-            'message' => 'data berhasil dihapus'
+            'message' => 'data berhasil dihapus',
+            'jwt' => $jwtDecoded
         ]);
     }
 
-    public function getWithToken(Request $request) {
+    public function getWithToken(Request $request)
+    {
         $mahasiswa = Mahasiswa::find($request->user->NIM);
         return response()->json([
             'status' => true,
